criando classe Midia e estebalecendo heranca
